refactor: 전체적인 재구조화, 모듈화 수행
