+++ core/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""
+Emoseum Core - 핵심 시스템 모듈
+"""
+
+from core.therapy_system import EmotionalImageTherapySystem
+
+__all__ = ["EmotionalImageTherapySystem"]
