+++ main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-감정 기반 디지털 치료 이미지 생성 시스템
+감정 기반 디지털 치료 이미지 생성 시스템 - CLI 인터페이스
 - SD-1.5 기반 경량화 이미지 생성
 - VAD 모델 기반 완벽한 감정 분석
 - LoRA 개인화 어댑터
@@ -8,46 +8,14 @@
 - CLI 기반 터미널 인터페이스
 """
 
-import os
 import sys
-import json
 import argparse
-import warnings
-import sqlite3
-import pickle
 import logging
 from pathlib import Path
-from datetime import datetime, timedelta
-from typing import Dict, List, Tuple, Optional, Any, Union
-from dataclasses import dataclass, asdict
-import math
-import random
 
-# 기본 라이브러리
-import numpy as np
-from PIL import Image
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-from torch.utils.data import DataLoader, Dataset
-
-# 경고 메시지 억제
-warnings.filterwarnings("ignore", category=FutureWarning)
-warnings.filterwarnings("ignore", category=UserWarning)
-
-# Apple 최적화 설정
-if torch.backends.mps.is_available():
-    device_type = "mps"
-    torch.mps.set_per_process_memory_fraction(0.8)
-elif torch.cuda.is_available():
-    device_type = "cuda"
-    torch.backends.cudnn.benchmark = True
-else:
-    device_type = "cpu"
-
-device = torch.device(device_type)
-print(f"🔧 디바이스: {device}")
+# 모듈 임포트
+from config import device, logger
+from core.therapy_system import EmotionalImageTherapySystem
 
 # 로깅 설정
 logging.basicConfig(
@@ -58,1924 +26,150 @@ logging.basicConfig(
         logging.StreamHandler(sys.stdout),
     ],
 )
-logger = logging.getLogger(__name__)
 
-# 필수 라이브러리 동적 임포트
-try:
-    from transformers import (
-        CLIPTextModel,
-        CLIPTokenizer,
-        AutoTokenizer,
-        AutoModel,
-        pipeline,
-        RobertaTokenizer,
-        RobertaModel,
+
+def check_system_requirements():
+    """시스템 요구사항 및 라이브러리 설치 상태 확인"""
+    import torch
+    import numpy as np
+    from emoseum.config import (
+        TRANSFORMERS_AVAILABLE,
+        DIFFUSERS_AVAILABLE,
+        PEFT_AVAILABLE,
     )
 
-    TRANSFORMERS_AVAILABLE = True
-    logger.info("✅ Transformers 라이브러리 로드 완료")
-except ImportError:
-    logger.error("❌ transformers 라이브러리가 필요합니다: pip install transformers")
-    TRANSFORMERS_AVAILABLE = False
+    print("🔍 시스템 요구사항 확인")
+    print("=" * 50)
 
-try:
-    from diffusers import (
-        StableDiffusionPipeline,
-        UNet2DConditionModel,
-        DDPMScheduler,
-        AutoencoderKL,
-        DPMSolverMultistepScheduler,
-        EulerDiscreteScheduler,
+    # Python 버전
+    python_version = sys.version_info
+    print(
+        f"🐍 Python: {python_version.major}.{python_version.minor}.{python_version.micro}"
     )
 
-    DIFFUSERS_AVAILABLE = True
-    logger.info("✅ Diffusers 라이브러리 로드 완료")
-except ImportError:
-    logger.error("❌ diffusers 라이브러리가 필요합니다: pip install diffusers")
-    DIFFUSERS_AVAILABLE = False
+    # 디바이스 정보
+    print(f"🔧 디바이스: {device} ({device.type})")
 
-try:
-    from peft import LoraConfig, get_peft_model, TaskType, PeftModel
+    if device.type == "mps":
+        print("🍎 Apple Silicon 최적화 활성화")
+    elif device.type == "cuda":
+        print(f"🚀 CUDA 가능 (GPU: {torch.cuda.get_device_name()})")
+    else:
+        print("💻 CPU 모드")
 
-    PEFT_AVAILABLE = True
-    logger.info("✅ PEFT 라이브러리 로드 완료")
-except ImportError:
-    logger.error("❌ peft 라이브러리가 필요합니다: pip install peft")
-    PEFT_AVAILABLE = False
+    # 메모리 정보
+    if device.type == "mps":
+        print("💾 통합 메모리 (Apple Silicon)")
+    elif device.type == "cuda":
+        gpu_memory = torch.cuda.get_device_properties(0).total_memory / 1024**3
+        print(f"💾 GPU 메모리: {gpu_memory:.1f}GB")
 
-# =============================================================================
-# 감정 임베딩 및 VAD 모델
-# =============================================================================
+    # 라이브러리 상태
+    print("\n📚 라이브러리 상태:")
+    libraries = {
+        "PyTorch": torch.__version__,
+        "Transformers": TRANSFORMERS_AVAILABLE,
+        "Diffusers": DIFFUSERS_AVAILABLE,
+        "PEFT": PEFT_AVAILABLE,
+        "PIL": True,  # 기본 라이브러리
+        "NumPy": np.__version__,
+    }
 
-
-@dataclass
-class EmotionEmbedding:
-    """Valence-Arousal-Dominance 기반 감정 임베딩"""
-
-    valence: float  # -1.0 (부정) to 1.0 (긍정)
-    arousal: float  # -1.0 (차분) to 1.0 (흥분)
-    dominance: float = 0.0  # -1.0 (수동) to 1.0 (지배적)
-    confidence: float = 1.0  # 감정 예측 신뢰도
-
-    def to_vector(self) -> np.ndarray:
-        return np.array([self.valence, self.arousal, self.dominance])
-
-    def to_dict(self) -> Dict[str, float]:
-        return asdict(self)
-
-    @classmethod
-    def from_vector(cls, vector: np.ndarray, confidence: float = 1.0):
-        return cls(
-            valence=float(vector[0]),
-            arousal=float(vector[1]),
-            dominance=float(vector[2]) if len(vector) > 2 else 0.0,
-            confidence=confidence,
-        )
-
-    def distance_to(self, other: "EmotionEmbedding") -> float:
-        """다른 감정과의 유클리드 거리"""
-        return np.linalg.norm(self.to_vector() - other.to_vector())
-
-    def similarity_to(self, other: "EmotionEmbedding") -> float:
-        """다른 감정과의 유사도 (0-1)"""
-        max_distance = np.sqrt(3 * 4)  # 최대 거리 (각 차원 -2 to 2)
-        return 1.0 - (self.distance_to(other) / max_distance)
-
-
-class AdvancedEmotionMapper:
-    """고급 VAD 기반 감정 매핑 시스템"""
-
-    def __init__(self, model_name="klue/roberta-large"):
-        self.device = device
-        self.model_name = model_name
-
-        # 감정 어휘 사전 (한국어 + 영어)
-        self.emotion_lexicon = {
-            # 기본 감정들
-            "기쁨": EmotionEmbedding(0.8, 0.6, 0.4),
-            "행복": EmotionEmbedding(0.8, 0.5, 0.3),
-            "즐거움": EmotionEmbedding(0.7, 0.7, 0.4),
-            "신남": EmotionEmbedding(0.9, 0.8, 0.6),
-            "만족": EmotionEmbedding(0.6, 0.2, 0.3),
-            "뿌듯": EmotionEmbedding(0.7, 0.4, 0.7),
-            "슬픔": EmotionEmbedding(-0.7, -0.3, -0.5),
-            "우울": EmotionEmbedding(-0.8, -0.4, -0.6),
-            "허무": EmotionEmbedding(-0.5, -0.6, -0.4),
-            "절망": EmotionEmbedding(-0.9, 0.3, -0.8),
-            "상실": EmotionEmbedding(-0.8, -0.2, -0.6),
-            "외로움": EmotionEmbedding(-0.6, -0.2, -0.6),
-            "화남": EmotionEmbedding(-0.6, 0.8, 0.7),
-            "분노": EmotionEmbedding(-0.8, 0.9, 0.8),
-            "짜증": EmotionEmbedding(-0.5, 0.6, 0.4),
-            "답답": EmotionEmbedding(-0.4, 0.5, -0.2),
-            "억울": EmotionEmbedding(-0.7, 0.6, -0.3),
-            "두려움": EmotionEmbedding(-0.8, 0.7, -0.8),
-            "걱정": EmotionEmbedding(-0.5, 0.6, -0.4),
-            "불안": EmotionEmbedding(-0.5, 0.7, -0.5),
-            "무서움": EmotionEmbedding(-0.8, 0.8, -0.7),
-            "긴장": EmotionEmbedding(-0.2, 0.8, -0.3),
-            "놀람": EmotionEmbedding(0.2, 0.9, 0.1),
-            "당황": EmotionEmbedding(-0.2, 0.8, -0.4),
-            "충격": EmotionEmbedding(-0.3, 0.9, -0.2),
-            "평온": EmotionEmbedding(0.4, -0.7, 0.2),
-            "차분": EmotionEmbedding(0.3, -0.8, 0.1),
-            "편안": EmotionEmbedding(0.6, -0.5, 0.3),
-            "안정": EmotionEmbedding(0.5, -0.6, 0.4),
-            "스트레스": EmotionEmbedding(-0.6, 0.7, -0.3),
-            "피곤": EmotionEmbedding(-0.3, -0.8, -0.4),
-            "지침": EmotionEmbedding(-0.4, -0.7, -0.5),
-            "권태": EmotionEmbedding(-0.2, -0.8, -0.3),
-            "사랑": EmotionEmbedding(0.9, 0.5, 0.3),
-            "애정": EmotionEmbedding(0.8, 0.4, 0.4),
-            "그리움": EmotionEmbedding(0.3, 0.3, -0.2),
-            "감사": EmotionEmbedding(0.8, 0.3, 0.3),
-            "고마움": EmotionEmbedding(0.7, 0.2, 0.2),
-            # 영어 감정들
-            "joy": EmotionEmbedding(0.8, 0.6, 0.4),
-            "happiness": EmotionEmbedding(0.8, 0.5, 0.3),
-            "sadness": EmotionEmbedding(-0.7, -0.3, -0.5),
-            "anger": EmotionEmbedding(-0.6, 0.8, 0.7),
-            "fear": EmotionEmbedding(-0.8, 0.7, -0.8),
-            "surprise": EmotionEmbedding(0.2, 0.9, 0.1),
-            "love": EmotionEmbedding(0.9, 0.5, 0.3),
-            "peace": EmotionEmbedding(0.4, -0.7, 0.2),
-            "stress": EmotionEmbedding(-0.6, 0.7, -0.3),
-            "tired": EmotionEmbedding(-0.3, -0.8, -0.4),
-        }
-
-        # 감정 강화 표현들
-        self.emotion_intensifiers = {
-            "매우": 1.3,
-            "정말": 1.2,
-            "엄청": 1.4,
-            "너무": 1.3,
-            "완전": 1.4,
-            "조금": 0.7,
-            "약간": 0.6,
-            "살짝": 0.5,
-            "좀": 0.7,
-            "extremely": 1.4,
-            "very": 1.3,
-            "really": 1.2,
-            "quite": 1.1,
-            "slightly": 0.6,
-            "somewhat": 0.7,
-            "a bit": 0.6,
-        }
-
-        # 부정 표현들
-        self.negation_words = {
-            "안",
-            "못",
-            "없",
-            "아니",
-            "not",
-            "no",
-            "never",
-            "don't",
-            "can't",
-            "won't",
-        }
-
-        # Transformer 모델 로드
-        self.use_transformer = TRANSFORMERS_AVAILABLE
-        if self.use_transformer:
-            try:
-                self.tokenizer = AutoTokenizer.from_pretrained(model_name)
-                self.text_encoder = AutoModel.from_pretrained(model_name).to(
-                    self.device
-                )
-                self.text_encoder.eval()
-
-                # VAD 예측 헤드
-                hidden_size = self.text_encoder.config.hidden_size
-                self.vad_predictor = nn.Sequential(
-                    nn.Linear(hidden_size, 512),
-                    nn.ReLU(),
-                    nn.Dropout(0.3),
-                    nn.Linear(512, 256),
-                    nn.ReLU(),
-                    nn.Dropout(0.3),
-                    nn.Linear(256, 128),
-                    nn.ReLU(),
-                    nn.Linear(128, 3),  # valence, arousal, dominance
-                    nn.Tanh(),
-                ).to(self.device)
-
-                # 간단한 가중치 초기화
-                self._init_vad_predictor()
-
-                logger.info(f"✅ 고급 감정 분석 모델 로드 완료: {model_name}")
-            except Exception as e:
-                logger.warning(
-                    f"⚠️ Transformer 모델 로드 실패: {e}, 규칙 기반 시스템 사용"
-                )
-                self.use_transformer = False
-
-    def _init_vad_predictor(self):
-        """VAD 예측기 초기화"""
-        for module in self.vad_predictor:
-            if isinstance(module, nn.Linear):
-                nn.init.xavier_uniform_(module.weight)
-                nn.init.zeros_(module.bias)
-
-    def extract_emotion_from_text(self, text: str) -> EmotionEmbedding:
-        """텍스트에서 감정 추출 (다중 방법론 융합)"""
-        # 1. 규칙 기반 감정 분석
-        rule_based_emotion = self._rule_based_emotion_analysis(text)
-
-        # 2. Transformer 기반 분석 (가능한 경우)
-        if self.use_transformer:
-            try:
-                transformer_emotion = self._transformer_emotion_analysis(text)
-                # 두 결과를 가중 평균
-                final_emotion = self._combine_emotions(
-                    rule_based_emotion,
-                    transformer_emotion,
-                    rule_weight=0.4,
-                    transformer_weight=0.6,
-                )
-            except Exception as e:
-                logger.warning(f"Transformer 분석 실패: {e}, 규칙 기반 결과 사용")
-                final_emotion = rule_based_emotion
+    for lib, status in libraries.items():
+        if isinstance(status, bool):
+            status_str = "✅ 설치됨" if status else "❌ 미설치"
         else:
-            final_emotion = rule_based_emotion
+            status_str = f"✅ v{status}"
+        print(f"  • {lib}: {status_str}")
 
-        # 3. 후처리 및 정규화
-        final_emotion = self._post_process_emotion(final_emotion, text)
+    # 설치 권장사항
+    missing_libs = []
+    if not TRANSFORMERS_AVAILABLE:
+        missing_libs.append("transformers")
+    if not DIFFUSERS_AVAILABLE:
+        missing_libs.append("diffusers")
+    if not PEFT_AVAILABLE:
+        missing_libs.append("peft")
 
-        logger.info(
-            f"감정 분석 결과: V={final_emotion.valence:.3f}, A={final_emotion.arousal:.3f}, D={final_emotion.dominance:.3f}"
-        )
-        return final_emotion
+    if missing_libs:
+        print(f"\n⚠️ 누락된 라이브러리: {', '.join(missing_libs)}")
+        print("설치 명령어:")
+        print(f"pip install {' '.join(missing_libs)}")
+    else:
+        print("\n✅ 모든 필수 라이브러리가 설치되어 있습니다!")
 
-    def _rule_based_emotion_analysis(self, text: str) -> EmotionEmbedding:
-        """규칙 기반 감정 분석"""
-        text_lower = text.lower()
-        words = text_lower.split()
-
-        detected_emotions = []
-        emotion_weights = []
-
-        # 감정 단어 탐지
-        for i, word in enumerate(words):
-            # 감정 어휘 매칭
-            for emotion_word, emotion_emb in self.emotion_lexicon.items():
-                if emotion_word in word or word in emotion_word:
-                    # 강화 표현 체크
-                    intensity = 1.0
-                    if i > 0 and words[i - 1] in self.emotion_intensifiers:
-                        intensity = self.emotion_intensifiers[words[i - 1]]
-
-                    # 부정 표현 체크
-                    negated = False
-                    for j in range(max(0, i - 2), i):
-                        if words[j] in self.negation_words:
-                            negated = True
-                            break
-
-                    # 감정 임베딩 조정
-                    adjusted_emotion = EmotionEmbedding(
-                        valence=emotion_emb.valence
-                        * intensity
-                        * (-1 if negated else 1),
-                        arousal=emotion_emb.arousal * intensity,
-                        dominance=emotion_emb.dominance
-                        * intensity
-                        * (-1 if negated else 1),
-                    )
-
-                    detected_emotions.append(adjusted_emotion)
-                    emotion_weights.append(intensity)
-
-        if not detected_emotions:
-            # 기본 중성 감정
-            return EmotionEmbedding(0.0, 0.0, 0.0, confidence=0.3)
-
-        # 가중 평균 계산
-        total_weight = sum(emotion_weights)
-        avg_valence = (
-            sum(e.valence * w for e, w in zip(detected_emotions, emotion_weights))
-            / total_weight
-        )
-        avg_arousal = (
-            sum(e.arousal * w for e, w in zip(detected_emotions, emotion_weights))
-            / total_weight
-        )
-        avg_dominance = (
-            sum(e.dominance * w for e, w in zip(detected_emotions, emotion_weights))
-            / total_weight
-        )
-
-        confidence = min(
-            1.0, len(detected_emotions) / 3.0
-        )  # 감정 단어 개수 기반 신뢰도
-
-        return EmotionEmbedding(avg_valence, avg_arousal, avg_dominance, confidence)
-
-    def _transformer_emotion_analysis(self, text: str) -> EmotionEmbedding:
-        """Transformer 기반 감정 분석"""
-        if not self.use_transformer:
-            return EmotionEmbedding(0.0, 0.0, 0.0, confidence=0.0)
-
-        # 토큰화 및 인코딩
-        inputs = self.tokenizer(
-            text, return_tensors="pt", max_length=512, truncation=True, padding=True
-        ).to(self.device)
-
-        with torch.no_grad():
-            # 텍스트 특성 추출
-            outputs = self.text_encoder(**inputs)
-            # CLS 토큰 또는 평균 풀링 사용
-            if hasattr(outputs, "pooler_output") and outputs.pooler_output is not None:
-                text_features = outputs.pooler_output
-            else:
-                text_features = outputs.last_hidden_state.mean(dim=1)
-
-            # VAD 예측
-            vad_scores = self.vad_predictor(text_features)
-            vad_scores = vad_scores.squeeze().cpu().numpy()
-
-        return EmotionEmbedding(
-            valence=float(vad_scores[0]),
-            arousal=float(vad_scores[1]),
-            dominance=float(vad_scores[2]),
-            confidence=0.8,
-        )
-
-    def _combine_emotions(
-        self,
-        emotion1: EmotionEmbedding,
-        emotion2: EmotionEmbedding,
-        rule_weight: float = 0.5,
-        transformer_weight: float = 0.5,
-    ) -> EmotionEmbedding:
-        """두 감정 분석 결과를 결합"""
-        total_weight = rule_weight + transformer_weight
-        rule_weight /= total_weight
-        transformer_weight /= total_weight
-
-        # 신뢰도 기반 가중치 조정
-        rule_conf_weight = rule_weight * emotion1.confidence
-        trans_conf_weight = transformer_weight * emotion2.confidence
-        total_conf_weight = rule_conf_weight + trans_conf_weight
-
-        if total_conf_weight > 0:
-            rule_conf_weight /= total_conf_weight
-            trans_conf_weight /= total_conf_weight
+    # 디렉토리 상태
+    print(f"\n📁 작업 디렉토리:")
+    dirs_to_check = ["generated_images", "user_loras"]
+    for dir_name in dirs_to_check:
+        dir_path = Path(dir_name)
+        if dir_path.exists():
+            file_count = len(list(dir_path.glob("*")))
+            print(f"  • {dir_name}/: ✅ ({file_count}개 파일)")
         else:
-            rule_conf_weight = trans_conf_weight = 0.5
-
-        return EmotionEmbedding(
-            valence=emotion1.valence * rule_conf_weight
-            + emotion2.valence * trans_conf_weight,
-            arousal=emotion1.arousal * rule_conf_weight
-            + emotion2.arousal * trans_conf_weight,
-            dominance=emotion1.dominance * rule_conf_weight
-            + emotion2.dominance * trans_conf_weight,
-            confidence=min(1.0, emotion1.confidence + emotion2.confidence),
-        )
-
-    def _post_process_emotion(
-        self, emotion: EmotionEmbedding, text: str
-    ) -> EmotionEmbedding:
-        """감정 후처리 및 정규화"""
-        # 범위 제한
-        valence = np.clip(emotion.valence, -1.0, 1.0)
-        arousal = np.clip(emotion.arousal, -1.0, 1.0)
-        dominance = np.clip(emotion.dominance, -1.0, 1.0)
-
-        # 텍스트 길이 기반 신뢰도 조정
-        text_length_factor = min(1.0, len(text.split()) / 10.0)
-        confidence = emotion.confidence * text_length_factor
-
-        return EmotionEmbedding(valence, arousal, dominance, confidence)
-
-    def emotion_to_prompt_modifiers(self, emotion: EmotionEmbedding) -> str:
-        """감정을 프롬프트 수정자로 변환"""
-        modifiers = []
-
-        # Valence 기반 수정자
-        if emotion.valence > 0.6:
-            modifiers.extend(["bright", "cheerful", "uplifting", "positive"])
-        elif emotion.valence > 0.2:
-            modifiers.extend(["pleasant", "mild", "gentle"])
-        elif emotion.valence < -0.6:
-            modifiers.extend(["dark", "melancholic", "somber", "moody"])
-        elif emotion.valence < -0.2:
-            modifiers.extend(["subdued", "quiet", "contemplative"])
-
-        # Arousal 기반 수정자
-        if emotion.arousal > 0.6:
-            modifiers.extend(["dynamic", "energetic", "vibrant", "intense"])
-        elif emotion.arousal > 0.2:
-            modifiers.extend(["lively", "animated"])
-        elif emotion.arousal < -0.6:
-            modifiers.extend(["calm", "peaceful", "serene", "tranquil"])
-        elif emotion.arousal < -0.2:
-            modifiers.extend(["relaxed", "soft"])
-
-        # Dominance 기반 수정자
-        if emotion.dominance > 0.4:
-            modifiers.extend(["bold", "confident", "strong"])
-        elif emotion.dominance < -0.4:
-            modifiers.extend(["delicate", "subtle", "gentle"])
-
-        # 감정 강도 기반 수정자
-        intensity = np.sqrt(
-            emotion.valence**2 + emotion.arousal**2 + emotion.dominance**2
-        ) / np.sqrt(3)
-        if intensity > 0.8:
-            modifiers.append("highly detailed")
-        elif intensity < 0.3:
-            modifiers.append("minimalist")
-
-        return ", ".join(modifiers[:6])  # 최대 6개 수정자
-
-
-# =============================================================================
-# LoRA 개인화 시스템
-# =============================================================================
-
-
-class PersonalizedLoRAManager:
-    """개인화된 LoRA 어댑터 관리 시스템"""
-
-    def __init__(
-        self,
-        base_model_path: str = "runwayml/stable-diffusion-v1-5",
-        lora_rank: int = 16,
-    ):
-        self.base_model_path = base_model_path
-        self.lora_rank = lora_rank
-        self.device = device
-        self.user_adapters = {}
-        self.adapter_configs = {}
-
-        # LoRA 저장 경로
-        self.lora_dir = Path("user_loras")
-        self.lora_dir.mkdir(exist_ok=True)
-
-        if not PEFT_AVAILABLE:
-            logger.warning("⚠️ PEFT 라이브러리가 없어 LoRA 기능이 제한됩니다")
-
-    def create_user_lora_config(self, user_id: str) -> Optional[LoraConfig]:
-        """사용자별 LoRA 설정 생성"""
-        if not PEFT_AVAILABLE:
-            return None
-
-        try:
-            lora_config = LoraConfig(
-                r=self.lora_rank,
-                lora_alpha=32,
-                target_modules=[
-                    "to_k",
-                    "to_q",
-                    "to_v",
-                    "to_out.0",
-                    "proj_in",
-                    "proj_out",
-                    "ff.net.0.proj",
-                    "ff.net.2",
-                ],
-                lora_dropout=0.1,
-                bias="none",
-                task_type=TaskType.DIFFUSION,
-            )
-
-            self.adapter_configs[user_id] = lora_config
-            logger.info(f"✅ 사용자 {user_id}의 LoRA 설정 생성 완료")
-            return lora_config
-
-        except Exception as e:
-            logger.error(f"❌ LoRA 설정 생성 실패: {e}")
-            return None
-
-    def save_user_lora(self, user_id: str, model_state_dict: Dict[str, torch.Tensor]):
-        """사용자 LoRA 어댑터 저장"""
-        try:
-            user_lora_path = self.lora_dir / f"{user_id}_lora.pt"
-            torch.save(model_state_dict, user_lora_path)
-            logger.info(f"✅ 사용자 {user_id} LoRA 저장: {user_lora_path}")
-        except Exception as e:
-            logger.error(f"❌ LoRA 저장 실패: {e}")
-
-    def load_user_lora(self, user_id: str) -> Optional[Dict[str, torch.Tensor]]:
-        """사용자 LoRA 어댑터 로드"""
-        try:
-            user_lora_path = self.lora_dir / f"{user_id}_lora.pt"
-            if user_lora_path.exists():
-                state_dict = torch.load(user_lora_path, map_location=self.device)
-                logger.info(f"✅ 사용자 {user_id} LoRA 로드: {user_lora_path}")
-                return state_dict
-        except Exception as e:
-            logger.error(f"❌ LoRA 로드 실패: {e}")
-        return None
-
-    def get_user_adapter_info(self, user_id: str) -> Dict[str, Any]:
-        """사용자 어댑터 정보 반환"""
-        return {
-            "user_id": user_id,
-            "lora_rank": self.lora_rank,
-            "config": self.adapter_configs.get(user_id),
-            "path": self.lora_dir / f"{user_id}_lora.pt",
-        }
-
-
-# =============================================================================
-# 사용자 프로파일 관리
-# =============================================================================
-
-
-class UserEmotionProfile:
-    """사용자 감정 프로파일 및 개인화 데이터 관리"""
-
-    def __init__(self, user_id: str, db_path: str = "user_profiles.db"):
-        self.user_id = user_id
-        self.db_path = db_path
-        self.emotion_history: List[Dict] = []
-        self.feedback_history: List[Dict] = []
-
-        # 개인화 선호도 가중치
-        self.preference_weights = {
-            "color_temperature": 0.0,  # -1.0 (차가운) to 1.0 (따뜻한)
-            "brightness": 0.0,  # -1.0 (어두운) to 1.0 (밝은)
-            "saturation": 0.0,  # -1.0 (무채색) to 1.0 (채도 높은)
-            "contrast": 0.0,  # -1.0 (낮은 대비) to 1.0 (높은 대비)
-            "complexity": 0.0,  # -1.0 (단순) to 1.0 (복잡)
-            "art_style": "realistic",  # realistic, abstract, impressionist, minimalist
-            "composition": "balanced",  # minimal, balanced, complex
-        }
-
-        # 치료 진행도 지표
-        self.therapeutic_progress = {
-            "mood_trend": 0.0,  # 감정 변화 트렌드
-            "stability_score": 0.0,  # 감정 안정성
-            "engagement_level": 0.0,  # 참여도
-            "recovery_indicator": 0.0,  # 회복 지표
-        }
-
-        # 학습 메타데이터
-        self.learning_metadata = {
-            "total_interactions": 0,
-            "positive_feedback_rate": 0.0,
-            "last_training_date": None,
-            "model_version": 1,
-        }
-
-        self._init_database()
-        self._load_profile()
-
-    def _init_database(self):
-        """SQLite 데이터베이스 초기화"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        # 감정 히스토리 테이블
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS emotion_history (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                user_id TEXT NOT NULL,
-                timestamp TEXT NOT NULL,
-                input_text TEXT,
-                valence REAL,
-                arousal REAL,
-                dominance REAL,
-                confidence REAL,
-                generated_prompt TEXT,
-                image_path TEXT
-            )
-        """
-        )
-
-        # 피드백 히스토리 테이블
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS feedback_history (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                user_id TEXT NOT NULL,
-                emotion_id INTEGER,
-                timestamp TEXT NOT NULL,
-                feedback_score REAL,
-                feedback_type TEXT,
-                comments TEXT,
-                FOREIGN KEY (emotion_id) REFERENCES emotion_history (id)
-            )
-        """
-        )
-
-        # 사용자 프로파일 테이블
-        cursor.execute(
-            """
-            CREATE TABLE IF NOT EXISTS user_profiles (
-                user_id TEXT PRIMARY KEY,
-                preference_weights TEXT,
-                therapeutic_progress TEXT,
-                learning_metadata TEXT,
-                last_updated TEXT
-            )
-        """
-        )
-
-        conn.commit()
-        conn.close()
-        logger.info(f"✅ 사용자 {self.user_id} 데이터베이스 초기화 완료")
-
-    def _load_profile(self):
-        """프로파일 데이터 로드"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        try:
-            # 감정 히스토리 로드 (최근 50개)
-            cursor.execute(
-                """
-                SELECT * FROM emotion_history 
-                WHERE user_id = ? 
-                ORDER BY timestamp DESC 
-                LIMIT 50
-            """,
-                (self.user_id,),
-            )
-
-            emotion_rows = cursor.fetchall()
-            for row in emotion_rows:
-                self.emotion_history.append(
-                    {
-                        "id": row[0],
-                        "timestamp": row[2],
-                        "input_text": row[3],
-                        "emotion": EmotionEmbedding(row[4], row[5], row[6], row[7]),
-                        "generated_prompt": row[8],
-                        "image_path": row[9],
-                    }
-                )
-
-            # 피드백 히스토리 로드
-            cursor.execute(
-                """
-                SELECT * FROM feedback_history 
-                WHERE user_id = ? 
-                ORDER BY timestamp DESC 
-                LIMIT 50
-            """,
-                (self.user_id,),
-            )
-
-            feedback_rows = cursor.fetchall()
-            for row in feedback_rows:
-                self.feedback_history.append(
-                    {
-                        "id": row[0],
-                        "emotion_id": row[2],
-                        "timestamp": row[3],
-                        "feedback_score": row[4],
-                        "feedback_type": row[5],
-                        "comments": row[6],
-                    }
-                )
-
-            # 프로파일 설정 로드
-            cursor.execute(
-                "SELECT * FROM user_profiles WHERE user_id = ?", (self.user_id,)
-            )
-            profile_row = cursor.fetchone()
-
-            if profile_row:
-                self.preference_weights.update(json.loads(profile_row[1]))
-                self.therapeutic_progress.update(json.loads(profile_row[2]))
-                self.learning_metadata.update(json.loads(profile_row[3]))
-
-            logger.info(
-                f"✅ 사용자 {self.user_id} 프로파일 로드: 감정 {len(self.emotion_history)}개, 피드백 {len(self.feedback_history)}개"
-            )
-
-        except Exception as e:
-            logger.error(f"❌ 프로파일 로드 실패: {e}")
-        finally:
-            conn.close()
-
-    def add_emotion_record(
-        self,
-        input_text: str,
-        emotion: EmotionEmbedding,
-        generated_prompt: str,
-        image_path: str = None,
-    ) -> int:
-        """감정 기록 추가"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        try:
-            timestamp = datetime.now().isoformat()
-            cursor.execute(
-                """
-                INSERT INTO emotion_history 
-                (user_id, timestamp, input_text, valence, arousal, dominance, confidence, generated_prompt, image_path)
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-            """,
-                (
-                    self.user_id,
-                    timestamp,
-                    input_text,
-                    emotion.valence,
-                    emotion.arousal,
-                    emotion.dominance,
-                    emotion.confidence,
-                    generated_prompt,
-                    image_path,
-                ),
-            )
-
-            emotion_id = cursor.lastrowid
-            conn.commit()
-
-            # 메모리에도 추가
-            self.emotion_history.append(
-                {
-                    "id": emotion_id,
-                    "timestamp": timestamp,
-                    "input_text": input_text,
-                    "emotion": emotion,
-                    "generated_prompt": generated_prompt,
-                    "image_path": image_path,
-                }
-            )
-
-            logger.info(f"✅ 감정 기록 추가: ID {emotion_id}")
-            return emotion_id
-
-        except Exception as e:
-            logger.error(f"❌ 감정 기록 추가 실패: {e}")
-            return -1
-        finally:
-            conn.close()
-
-    def add_feedback(
-        self,
-        emotion_id: int,
-        feedback_score: float,
-        feedback_type: str = "rating",
-        comments: str = None,
-    ):
-        """피드백 추가"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        try:
-            timestamp = datetime.now().isoformat()
-            cursor.execute(
-                """
-                INSERT INTO feedback_history 
-                (user_id, emotion_id, timestamp, feedback_score, feedback_type, comments)
-                VALUES (?, ?, ?, ?, ?, ?)
-            """,
-                (
-                    self.user_id,
-                    emotion_id,
-                    timestamp,
-                    feedback_score,
-                    feedback_type,
-                    comments,
-                ),
-            )
-
-            conn.commit()
-
-            # 메모리에도 추가
-            self.feedback_history.append(
-                {
-                    "emotion_id": emotion_id,
-                    "timestamp": timestamp,
-                    "feedback_score": feedback_score,
-                    "feedback_type": feedback_type,
-                    "comments": comments,
-                }
-            )
-
-            # 개인화 선호도 업데이트
-            self._update_preferences_from_feedback(feedback_score)
-
-            # 치료 진행도 업데이트
-            self._update_therapeutic_progress()
-
-            # 프로파일 저장
-            self._save_profile()
-
-            logger.info(f"✅ 피드백 추가: 감정 ID {emotion_id}, 점수 {feedback_score}")
-
-        except Exception as e:
-            logger.error(f"❌ 피드백 추가 실패: {e}")
-        finally:
-            conn.close()
-
-    def _update_preferences_from_feedback(self, feedback_score: float):
-        """피드백 점수를 기반으로 선호도 업데이트"""
-        learning_rate = 0.1
-
-        if feedback_score > 3.0:  # 긍정적 피드백 (1-5 척도)
-            weight = (feedback_score - 3.0) / 2.0 * learning_rate
-
-            # 최근 감정 기반 선호도 조정
-            if self.emotion_history:
-                recent_emotion = self.emotion_history[-1]["emotion"]
-
-                # Valence 기반 밝기/채도 조정
-                if recent_emotion.valence > 0:
-                    self.preference_weights["brightness"] += weight * 0.1
-                    self.preference_weights["saturation"] += weight * 0.1
-                else:
-                    self.preference_weights["brightness"] -= weight * 0.05
-                    self.preference_weights["saturation"] -= weight * 0.05
-
-                # Arousal 기반 대비/복잡성 조정
-                if recent_emotion.arousal > 0:
-                    self.preference_weights["contrast"] += weight * 0.1
-                    self.preference_weights["complexity"] += weight * 0.05
-                else:
-                    self.preference_weights["contrast"] -= weight * 0.05
-                    self.preference_weights["complexity"] -= weight * 0.1
-
-        # 범위 제한
-        for key in self.preference_weights:
-            if isinstance(self.preference_weights[key], (int, float)):
-                self.preference_weights[key] = np.clip(
-                    self.preference_weights[key], -1.0, 1.0
-                )
-
-    def _update_therapeutic_progress(self):
-        """치료 진행도 업데이트"""
-        if len(self.emotion_history) < 3:
-            return
-
-        # 최근 감정들의 Valence 트렌드 분석
-        recent_valences = [
-            entry["emotion"].valence for entry in self.emotion_history[-10:]
-        ]
-        if len(recent_valences) >= 3:
-            # 선형 회귀로 트렌드 계산
-            x = np.arange(len(recent_valences))
-            y = np.array(recent_valences)
-            slope = np.corrcoef(x, y)[0, 1] if len(recent_valences) > 1 else 0
-            self.therapeutic_progress["mood_trend"] = slope
-
-        # 감정 안정성 (변동성의 역수)
-        if len(recent_valences) >= 5:
-            stability = 1.0 / (1.0 + np.std(recent_valences))
-            self.therapeutic_progress["stability_score"] = stability
-
-        # 참여도 (피드백 제공률)
-        if self.feedback_history:
-            recent_interactions = len(self.emotion_history[-20:])
-            recent_feedbacks = len(
-                [
-                    f
-                    for f in self.feedback_history[-20:]
-                    if f["emotion_id"] in [e["id"] for e in self.emotion_history[-20:]]
-                ]
-            )
-            engagement = recent_feedbacks / max(1, recent_interactions)
-            self.therapeutic_progress["engagement_level"] = engagement
-
-        # 회복 지표 (긍정적 피드백 비율 + 감정 트렌드)
-        if self.feedback_history:
-            positive_feedbacks = len(
-                [f for f in self.feedback_history[-20:] if f["feedback_score"] > 3.0]
-            )
-            total_feedbacks = len(self.feedback_history[-20:])
-            positive_rate = positive_feedbacks / max(1, total_feedbacks)
-
-            recovery = (
-                positive_rate
-                + max(0, self.therapeutic_progress["mood_trend"])
-                + self.therapeutic_progress["stability_score"]
-            ) / 3.0
-            self.therapeutic_progress["recovery_indicator"] = recovery
-
-    def _save_profile(self):
-        """프로파일 데이터 저장"""
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.cursor()
-
-        try:
-            timestamp = datetime.now().isoformat()
-            cursor.execute(
-                """
-                INSERT OR REPLACE INTO user_profiles 
-                (user_id, preference_weights, therapeutic_progress, learning_metadata, last_updated)
-                VALUES (?, ?, ?, ?, ?)
-            """,
-                (
-                    self.user_id,
-                    json.dumps(self.preference_weights),
-                    json.dumps(self.therapeutic_progress),
-                    json.dumps(self.learning_metadata),
-                    timestamp,
-                ),
-            )
-
-            conn.commit()
-            logger.info(f"✅ 사용자 {self.user_id} 프로파일 저장 완료")
-
-        except Exception as e:
-            logger.error(f"❌ 프로파일 저장 실패: {e}")
-        finally:
-            conn.close()
-
-    def get_personalized_style_modifiers(self) -> str:
-        """개인화된 스타일 수정자 생성"""
-        modifiers = []
-
-        # 색온도 기반
-        if self.preference_weights["color_temperature"] > 0.3:
-            modifiers.append("warm lighting")
-        elif self.preference_weights["color_temperature"] < -0.3:
-            modifiers.append("cool lighting")
-
-        # 밝기 기반
-        if self.preference_weights["brightness"] > 0.3:
-            modifiers.append("bright")
-        elif self.preference_weights["brightness"] < -0.3:
-            modifiers.append("dim lighting")
-
-        # 채도 기반
-        if self.preference_weights["saturation"] > 0.3:
-            modifiers.append("vibrant colors")
-        elif self.preference_weights["saturation"] < -0.3:
-            modifiers.append("muted colors")
-
-        # 대비 기반
-        if self.preference_weights["contrast"] > 0.3:
-            modifiers.append("high contrast")
-        elif self.preference_weights["contrast"] < -0.3:
-            modifiers.append("soft contrast")
-
-        # 복잡성 기반
-        if self.preference_weights["complexity"] > 0.3:
-            modifiers.append("detailed")
-        elif self.preference_weights["complexity"] < -0.3:
-            modifiers.append("minimalist")
-
-        # 아트 스타일
-        modifiers.append(f"{self.preference_weights['art_style']} style")
-
-        return ", ".join(modifiers)
-
-    def get_therapeutic_insights(self) -> Dict[str, Any]:
-        """치료적 인사이트 제공"""
-        if len(self.emotion_history) < 3:
-            return {
-                "message": "충분한 데이터가 수집되지 않았습니다.",
-                "status": "insufficient_data",
-            }
-
-        insights = {
-            "emotional_state": {
-                "current_mood": self._get_current_mood_description(),
-                "mood_trend": self.therapeutic_progress["mood_trend"],
-                "stability": self.therapeutic_progress["stability_score"],
-            },
-            "progress_indicators": {
-                "engagement_level": self.therapeutic_progress["engagement_level"],
-                "recovery_indicator": self.therapeutic_progress["recovery_indicator"],
-                "total_interactions": len(self.emotion_history),
-                "feedback_count": len(self.feedback_history),
-            },
-            "recommendations": self._generate_recommendations(),
-            "preference_summary": self.preference_weights,
-        }
-
-        return insights
-
-    def _get_current_mood_description(self) -> str:
-        """현재 기분 상태 설명"""
-        if not self.emotion_history:
-            return "데이터 없음"
-
-        recent_emotions = [entry["emotion"] for entry in self.emotion_history[-5:]]
-        avg_valence = np.mean([e.valence for e in recent_emotions])
-        avg_arousal = np.mean([e.arousal for e in recent_emotions])
-
-        if avg_valence > 0.3 and avg_arousal > 0.3:
-            return "활기찬 긍정 상태"
-        elif avg_valence > 0.3 and avg_arousal < -0.3:
-            return "평온한 긍정 상태"
-        elif avg_valence < -0.3 and avg_arousal > 0.3:
-            return "불안한 부정 상태"
-        elif avg_valence < -0.3 and avg_arousal < -0.3:
-            return "우울한 상태"
-        else:
-            return "중성적 상태"
-
-    def _generate_recommendations(self) -> List[str]:
-        """개인화된 추천사항 생성"""
-        recommendations = []
-
-        # 감정 트렌드 기반
-        if self.therapeutic_progress["mood_trend"] < -0.3:
-            recommendations.append(
-                "부정적인 감정 패턴이 감지되었습니다. 긍정적인 활동이나 이미지 생성을 시도해보세요."
-            )
-        elif self.therapeutic_progress["mood_trend"] > 0.3:
-            recommendations.append(
-                "감정 상태가 개선되고 있습니다. 현재 패턴을 유지하세요."
-            )
-
-        # 안정성 기반
-        if self.therapeutic_progress["stability_score"] < 0.5:
-            recommendations.append(
-                "감정 변동이 큽니다. 규칙적인 사용과 일관된 피드백이 도움될 것입니다."
-            )
-
-        # 참여도 기반
-        if self.therapeutic_progress["engagement_level"] < 0.3:
-            recommendations.append(
-                "더 자주 피드백을 제공하시면 개인화 효과가 향상됩니다."
-            )
-
-        # 기본 추천
-        if not recommendations:
-            recommendations.append(
-                "현재 상태가 양호합니다. 지속적인 사용을 권장합니다."
-            )
-
-        return recommendations
-
-
-# =============================================================================
-# DRaFT+ 강화학습 시스템
-# =============================================================================
-
-
-class DRaFTPlusRewardModel:
-    """DRaFT+ 방식의 개선된 보상 모델"""
-
-    def __init__(self, device: torch.device = None):
-        self.device = device if device else torch.device("cpu")
-
-        # 감정 정확도 평가기
-        self.emotion_evaluator = self._build_emotion_evaluator()
-
-        # 미적 품질 평가기
-        self.aesthetic_evaluator = self._build_aesthetic_evaluator()
-
-        # 개인화 점수 평가기
-        self.personalization_evaluator = self._build_personalization_evaluator()
-
-        # 다양성 평가기
-        self.diversity_evaluator = self._build_diversity_evaluator()
-
-        logger.info("✅ DRaFT+ 보상 모델 초기화 완료")
-
-    def _build_emotion_evaluator(self) -> nn.Module:
-        """감정 정확도 평가기"""
-        return nn.Sequential(
-            nn.Linear(768, 512),  # CLIP 임베딩 크기 가정
-            nn.ReLU(),
-            nn.Dropout(0.2),
-            nn.Linear(512, 256),
-            nn.ReLU(),
-            nn.Dropout(0.2),
-            nn.Linear(256, 3),  # VAD 예측
-            nn.Tanh(),
-        ).to(self.device)
-
-    def _build_aesthetic_evaluator(self) -> nn.Module:
-        """미적 품질 평가기"""
-        return nn.Sequential(
-            nn.Conv2d(3, 32, 3, padding=1),
-            nn.ReLU(),
-            nn.MaxPool2d(2),
-            nn.Conv2d(32, 64, 3, padding=1),
-            nn.ReLU(),
-            nn.MaxPool2d(2),
-            nn.Conv2d(64, 128, 3, padding=1),
-            nn.ReLU(),
-            nn.AdaptiveAvgPool2d((4, 4)),
-            nn.Flatten(),
-            nn.Linear(128 * 4 * 4, 256),
-            nn.ReLU(),
-            nn.Dropout(0.3),
-            nn.Linear(256, 1),
-            nn.Sigmoid(),
-        ).to(self.device)
-
-    def _build_personalization_evaluator(self) -> nn.Module:
-        """개인화 점수 평가기"""
-        return nn.Sequential(
-            nn.Linear(512 + 7, 256),  # 이미지 특성 + 개인화 선호도
-            nn.ReLU(),
-            nn.Dropout(0.2),
-            nn.Linear(256, 128),
-            nn.ReLU(),
-            nn.Linear(128, 1),
-            nn.Sigmoid(),
-        ).to(self.device)
-
-    def _build_diversity_evaluator(self) -> nn.Module:
-        """다양성 평가기"""
-        return nn.Sequential(
-            nn.Linear(512, 256),
-            nn.ReLU(),
-            nn.Linear(256, 128),
-            nn.ReLU(),
-            nn.Linear(128, 1),
-            nn.Sigmoid(),
-        ).to(self.device)
-
-    def calculate_comprehensive_reward(
-        self,
-        generated_images: torch.Tensor,
-        target_emotion: EmotionEmbedding,
-        user_profile: UserEmotionProfile,
-        previous_images: List[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        """종합적인 보상 계산 (DRaFT+ 방식)"""
-        batch_size = generated_images.shape[0]
-
-        try:
-            with torch.no_grad():
-                # 1. 감정 정확도 보상
-                emotion_reward = self._calculate_emotion_reward(
-                    generated_images, target_emotion
-                )
-
-                # 2. 미적 품질 보상
-                aesthetic_reward = self._calculate_aesthetic_reward(generated_images)
-
-                # 3. 개인화 보상
-                personalization_reward = self._calculate_personalization_reward(
-                    generated_images, user_profile
-                )
-
-                # 4. 다양성 보상 (DRaFT+ 추가 요소)
-                diversity_reward = self._calculate_diversity_reward(
-                    generated_images, previous_images
-                )
-
-                # 5. 가중 합계 (DRaFT+ 가중치)
-                total_reward = (
-                    0.35 * emotion_reward
-                    + 0.25 * aesthetic_reward
-                    + 0.25 * personalization_reward
-                    + 0.15 * diversity_reward
-                )
-
-                # 6. 정규화 및 스무딩
-                total_reward = torch.clamp(total_reward, 0.0, 1.0)
-
-            return total_reward
-
-        except Exception as e:
-            logger.warning(f"⚠️ 보상 계산 실패: {e}, 기본값 반환")
-            return torch.tensor([0.5] * batch_size, device=self.device)
-
-    def _calculate_emotion_reward(
-        self, images: torch.Tensor, target_emotion: EmotionEmbedding
-    ) -> torch.Tensor:
-        """감정 정확도 기반 보상"""
-        batch_size = images.shape[0]
-
-        # 간단한 이미지 특성 추출 (실제로는 CLIP 등 사용)
-        image_features = self._extract_simple_features(images)
-
-        # 목표 감정과의 일치도 계산
-        target_vector = torch.tensor(
-            [target_emotion.valence, target_emotion.arousal, target_emotion.dominance],
-            device=self.device,
-        ).repeat(batch_size, 1)
-
-        predicted_emotions = self.emotion_evaluator(image_features)
-        emotion_distance = F.mse_loss(
-            predicted_emotions, target_vector, reduction="none"
-        ).mean(dim=1)
-
-        # 거리를 보상으로 변환 (거리가 클수록 보상 낮음)
-        emotion_reward = torch.exp(-emotion_distance * 2.0)
-
-        return emotion_reward
-
-    def _calculate_aesthetic_reward(self, images: torch.Tensor) -> torch.Tensor:
-        """미적 품질 보상"""
-        # 이미지 크기 조정 (필요시)
-        if images.shape[-1] != 64:  # 예시 크기
-            images_resized = F.interpolate(
-                images, size=(64, 64), mode="bilinear", align_corners=False
-            )
-        else:
-            images_resized = images
-
-        aesthetic_scores = self.aesthetic_evaluator(images_resized).squeeze()
-
-        # 배치 차원 보장
-        if aesthetic_scores.dim() == 0:
-            aesthetic_scores = aesthetic_scores.unsqueeze(0)
-
-        return aesthetic_scores
-
-    def _calculate_personalization_reward(
-        self, images: torch.Tensor, user_profile: UserEmotionProfile
-    ) -> torch.Tensor:
-        """개인화 보상"""
-        batch_size = images.shape[0]
-
-        # 이미지 특성 추출
-        image_features = self._extract_simple_features(images)
-
-        # 사용자 선호도 벡터 생성
-        preference_vector = torch.tensor(
-            [
-                user_profile.preference_weights["color_temperature"],
-                user_profile.preference_weights["brightness"],
-                user_profile.preference_weights["saturation"],
-                user_profile.preference_weights["contrast"],
-                user_profile.preference_weights["complexity"],
-                (
-                    1.0
-                    if user_profile.preference_weights["art_style"] == "realistic"
-                    else 0.0
-                ),
-                (
-                    1.0
-                    if user_profile.preference_weights["composition"] == "balanced"
-                    else 0.0
-                ),
-            ],
-            device=self.device,
-        ).repeat(batch_size, 1)
-
-        # 개인화 특성과 결합
-        combined_features = torch.cat([image_features, preference_vector], dim=1)
-        personalization_scores = self.personalization_evaluator(
-            combined_features
-        ).squeeze()
-
-        if personalization_scores.dim() == 0:
-            personalization_scores = personalization_scores.unsqueeze(0)
-
-        return personalization_scores
-
-    def _calculate_diversity_reward(
-        self, images: torch.Tensor, previous_images: List[torch.Tensor] = None
-    ) -> torch.Tensor:
-        """다양성 보상 (DRaFT+ 핵심 요소)"""
-        batch_size = images.shape[0]
-
-        if previous_images is None or len(previous_images) == 0:
-            # 이전 이미지가 없으면 최대 다양성 보상
-            return torch.ones(batch_size, device=self.device)
-
-        # 현재 이미지 특성 추출
-        current_features = self._extract_simple_features(images)
-
-        # 이전 이미지들과의 거리 계산
-        diversity_scores = []
-
-        for img_features in current_features:
-            min_distance = float("inf")
-
-            for prev_img in previous_images[-5:]:  # 최근 5개 이미지와 비교
-                if prev_img.shape[0] == 1:  # 배치 크기 1인 경우
-                    prev_features = self._extract_simple_features(prev_img).squeeze(0)
-                    distance = F.pairwise_distance(
-                        img_features.unsqueeze(0), prev_features.unsqueeze(0)
-                    )
-                    min_distance = min(min_distance, distance.item())
-
-            # 거리 기반 다양성 점수 (거리가 클수록 다양성 높음)
-            diversity_score = min(1.0, min_distance / 10.0)  # 정규화
-            diversity_scores.append(diversity_score)
-
-        return torch.tensor(diversity_scores, device=self.device)
-
-    def _extract_simple_features(self, images: torch.Tensor) -> torch.Tensor:
-        """간단한 이미지 특성 추출"""
-        batch_size = images.shape[0]
-
-        # 기본적인 통계적 특성들
-        features = []
-
-        for i in range(batch_size):
-            img = images[i]
-
-            # 색상 통계
-            mean_rgb = img.mean(dim=[1, 2])  # RGB 평균
-            std_rgb = img.std(dim=[1, 2])  # RGB 표준편차
-
-            # 밝기 및 대비
-            gray = 0.299 * img[0] + 0.587 * img[1] + 0.114 * img[2]
-            brightness = gray.mean()
-            contrast = gray.std()
-
-            # 에지 밀도 (간단한 근사)
-            grad_x = torch.abs(gray[1:, :] - gray[:-1, :]).mean()
-            grad_y = torch.abs(gray[:, 1:] - gray[:, :-1]).mean()
-            edge_density = (grad_x + grad_y) / 2
-
-            # 특성 벡터 구성
-            feature_vector = torch.cat(
-                [
-                    mean_rgb,
-                    std_rgb,
-                    brightness.unsqueeze(0),
-                    contrast.unsqueeze(0),
-                    edge_density.unsqueeze(0),
-                ]
-            )
-
-            # 512차원으로 패딩 (실제로는 더 정교한 특성 추출 필요)
-            if feature_vector.shape[0] < 512:
-                padding = torch.zeros(512 - feature_vector.shape[0], device=self.device)
-                feature_vector = torch.cat([feature_vector, padding])
-
-            features.append(feature_vector[:512])
-
-        return torch.stack(features)
-
-
-class DRaFTPlusTrainer:
-    """DRaFT+ 기반 강화학습 트레이너"""
-
-    def __init__(
-        self, pipeline, reward_model: DRaFTPlusRewardModel, learning_rate: float = 1e-5
-    ):
-        self.pipeline = pipeline
-        self.reward_model = reward_model
-        self.device = device
-        self.learning_rate = learning_rate
-
-        # 옵티마이저 설정
-        if hasattr(pipeline, "unet") and hasattr(pipeline.unet, "parameters"):
-            self.optimizer = optim.AdamW(
-                pipeline.unet.parameters(),
-                lr=learning_rate,
-                weight_decay=0.01,
-                eps=1e-8,
-            )
-            self.scheduler = optim.lr_scheduler.CosineAnnealingLR(
-                self.optimizer, T_max=100, eta_min=learning_rate * 0.1
-            )
-            self.can_train = True
-            logger.info("✅ DRaFT+ 트레이너 초기화 완료")
-        else:
-            logger.warning("⚠️ UNet이 없어 시뮬레이션 모드로 실행")
-            self.can_train = False
-
-        # 다양성을 위한 이미지 히스토리
-        self.image_history = []
-        self.max_history_size = 10
-
-    def train_step(
-        self,
-        prompt: str,
-        target_emotion: EmotionEmbedding,
-        user_profile: UserEmotionProfile,
-        num_inference_steps: int = 8,
-        diversity_weight: float = 0.15,
-    ) -> Dict[str, float]:
-        """DRaFT+ 학습 스텝"""
-
-        if not self.can_train:
-            # 시뮬레이션 모드
-            return {
-                "emotion_reward": random.uniform(0.4, 0.8),
-                "aesthetic_reward": random.uniform(0.5, 0.9),
-                "personalization_reward": random.uniform(0.3, 0.7),
-                "diversity_reward": random.uniform(0.6, 1.0),
-                "total_reward": random.uniform(0.5, 0.8),
-                "loss": random.uniform(0.2, 0.6),
-                "learning_rate": self.learning_rate,
-                "mode": "simulation",
-            }
-
-        try:
-            # 그래디언트 초기화
-            self.optimizer.zero_grad()
-
-            # UNet 학습 모드로 설정
-            self.pipeline.unet.train()
-
-            # 이미지 생성 (간소화된 디퓨전 과정)
-            with torch.enable_grad():
-                # 텍스트 임베딩
-                text_embeddings = self._encode_prompt(prompt)
-
-                # 노이즈 생성
-                latents = torch.randn(
-                    (1, 4, 64, 64),  # SD 1.5 기본 latent 크기
-                    device=self.device,
-                    dtype=text_embeddings.dtype,
-                    requires_grad=True,
-                )
-
-                # 간소화된 디노이징 (빠른 학습을 위해)
-                for step in range(num_inference_steps):
-                    t = torch.tensor(
-                        [1000 - step * (1000 // num_inference_steps)],
-                        device=self.device,
-                    )
-
-                    # UNet 예측
-                    noise_pred = self.pipeline.unet(
-                        latents,
-                        t,
-                        encoder_hidden_states=text_embeddings,
-                        return_dict=False,
-                    )[0]
-
-                    # 디노이징 스텝
-                    latents = latents - 0.1 * noise_pred
-
-                # VAE 디코딩 (가능한 경우)
-                if hasattr(self.pipeline, "vae"):
-                    try:
-                        if hasattr(self.pipeline.vae.config, "scaling_factor"):
-                            latents_scaled = (
-                                latents / self.pipeline.vae.config.scaling_factor
-                            )
-                        else:
-                            latents_scaled = latents
-
-                        images = self.pipeline.vae.decode(
-                            latents_scaled, return_dict=False
-                        )[0]
-                        images = (images / 2 + 0.5).clamp(0, 1)
-                    except:
-                        # VAE 디코딩 실패시 가짜 이미지
-                        images = torch.rand(1, 3, 512, 512, device=self.device)
-                else:
-                    images = torch.rand(1, 3, 512, 512, device=self.device)
-
-                # 보상 계산
-                rewards = self.reward_model.calculate_comprehensive_reward(
-                    images, target_emotion, user_profile, self.image_history
-                )
-
-                # DRaFT+ 손실 계산 (다양성 정규화 포함)
-                reward_loss = -rewards.mean()
-
-                # 다양성 정규화 손실
-                diversity_loss = self._calculate_diversity_loss(images)
-
-                # 총 손실
-                total_loss = reward_loss + diversity_weight * diversity_loss
-
-                # 역전파
-                total_loss.backward()
-
-                # 그래디언트 클리핑
-                torch.nn.utils.clip_grad_norm_(self.pipeline.unet.parameters(), 1.0)
-
-                # 옵티마이저 스텝
-                self.optimizer.step()
-                self.scheduler.step()
-
-                # 이미지 히스토리 업데이트
-                self._update_image_history(images.detach())
-
-                # 결과 반환
-                with torch.no_grad():
-                    return {
-                        "total_reward": rewards.mean().item(),
-                        "reward_loss": reward_loss.item(),
-                        "diversity_loss": diversity_loss.item(),
-                        "total_loss": total_loss.item(),
-                        "learning_rate": self.scheduler.get_last_lr()[0],
-                        "mode": "training",
-                    }
-
-        except Exception as e:
-            logger.error(f"❌ DRaFT+ 학습 실패: {e}")
-            return {
-                "error": str(e),
-                "total_reward": 0.5,
-                "loss": 1.0,
-                "learning_rate": self.learning_rate,
-                "mode": "error",
-            }
-
-    def _encode_prompt(self, prompt: str) -> torch.Tensor:
-        """프롬프트 인코딩"""
-        try:
-            text_inputs = self.pipeline.tokenizer(
-                prompt,
-                padding="max_length",
-                max_length=self.pipeline.tokenizer.model_max_length,
-                truncation=True,
-                return_tensors="pt",
-            ).to(self.device)
-
-            text_embeddings = self.pipeline.text_encoder(text_inputs.input_ids)[0]
-            return text_embeddings
-
-        except Exception as e:
-            logger.warning(f"⚠️ 프롬프트 인코딩 실패: {e}, 기본값 사용")
-            return torch.randn(1, 77, 768, device=self.device)
-
-    def _calculate_diversity_loss(self, images: torch.Tensor) -> torch.Tensor:
-        """다양성 손실 계산 (DRaFT+ 핵심)"""
-        if len(self.image_history) == 0:
-            return torch.tensor(0.0, device=self.device)
-
-        # 현재 이미지와 히스토리 이미지들 간의 유사성 계산
-        current_features = self.reward_model._extract_simple_features(images)
-
-        total_similarity = 0.0
-        count = 0
-
-        for hist_img in self.image_history[-3:]:  # 최근 3개와 비교
-            hist_features = self.reward_model._extract_simple_features(hist_img)
-            similarity = F.cosine_similarity(
-                current_features, hist_features, dim=1
-            ).mean()
-            total_similarity += similarity
-            count += 1
-
-        if count > 0:
-            avg_similarity = total_similarity / count
-            # 유사성이 높을수록 다양성 손실 증가
-            diversity_loss = torch.clamp(avg_similarity, 0.0, 1.0)
-        else:
-            diversity_loss = torch.tensor(0.0, device=self.device)
-
-        return diversity_loss
-
-    def _update_image_history(self, images: torch.Tensor):
-        """이미지 히스토리 업데이트"""
-        self.image_history.append(images.clone())
-
-        # 히스토리 크기 제한
-        if len(self.image_history) > self.max_history_size:
-            self.image_history.pop(0)
-
-
-# =============================================================================
-# 메인 시스템 통합
-# =============================================================================
-
-
-class EmotionalImageTherapySystem:
-    """감정 기반 이미지 치료 시스템"""
-
-    def __init__(self, model_path: str = "runwayml/stable-diffusion-v1-5"):
-        self.model_path = model_path
-        self.device = device
-
-        # 출력 디렉토리 생성
-        self.output_dir = Path("generated_images")
-        self.output_dir.mkdir(exist_ok=True)
-
-        # 컴포넌트 초기화
-        logger.info("🚀 시스템 초기화 시작...")
-
-        # 1. 감정 매퍼 초기화
-        self.emotion_mapper = AdvancedEmotionMapper()
-
-        # 2. LoRA 매니저 초기화
-        self.lora_manager = PersonalizedLoRAManager(model_path)
-
-        # 3. SD 파이프라인 로드
-        self.pipeline = self._load_pipeline()
-
-        # 4. 보상 모델 및 트레이너 초기화
-        if self.pipeline:
-            self.reward_model = DRaFTPlusRewardModel(self.device)
-            self.trainer = DRaFTPlusTrainer(self.pipeline, self.reward_model)
-        else:
-            self.reward_model = None
-            self.trainer = None
-
-        # 5. 사용자 프로파일 캐시
-        self.user_profiles = {}
-
-        logger.info("✅ 시스템 초기화 완료!")
-
-    def _load_pipeline(self):
-        """SD 파이프라인 로드"""
-        if not DIFFUSERS_AVAILABLE:
-            logger.error("❌ Diffusers 라이브러리가 필요합니다")
-            return None
-
-        try:
-            logger.info(f"📦 Stable Diffusion 파이프라인 로드 중: {self.model_path}")
-
-            pipeline = StableDiffusionPipeline.from_pretrained(
-                self.model_path,
-                torch_dtype=(
-                    # 유효하지 않은 숫자가 들어가는 오류가 발생하므로 모두 float32로 설정
-                    torch.float32
-                    if self.device.type == "mps"
-                    else torch.float32
-                ),
-                use_safetensors=True,
-                safety_checker=None,  # 빠른 생성을 위해 비활성화
-                requires_safety_checker=False,
-            )
-
-            # 최적화 설정
-            pipeline = pipeline.to(self.device)
-
-            # 메모리 최적화
-            pipeline.enable_attention_slicing()
-
-            if self.device.type == "cuda":
-                pipeline.enable_sequential_cpu_offload()
-
-            # 빠른 스케줄러로 변경
-            pipeline.scheduler = EulerDiscreteScheduler.from_config(
-                pipeline.scheduler.config
-            )
-
-            logger.info("✅ SD 파이프라인 로드 및 최적화 완료")
-            return pipeline
-
-        except Exception as e:
-            logger.error(f"❌ SD 파이프라인 로드 실패: {e}")
-            return None
-
-    def get_user_profile(self, user_id: str) -> UserEmotionProfile:
-        """사용자 프로파일 가져오기 또는 생성"""
-        if user_id not in self.user_profiles:
-            self.user_profiles[user_id] = UserEmotionProfile(user_id)
-            logger.info(f"✅ 새 사용자 프로파일 생성: {user_id}")
-        return self.user_profiles[user_id]
-
-    def generate_therapeutic_image(
-        self,
-        user_id: str,
-        input_text: str,
-        base_prompt: str = "",
-        num_inference_steps: int = 15,
-        guidance_scale: float = 7.5,
-        width: int = 512,
-        height: int = 512,
-    ) -> Dict[str, Any]:
-        """치료용 이미지 생성"""
-
-        try:
-            logger.info(f"🎨 사용자 {user_id}의 이미지 생성 시작")
-            logger.info(f"📝 입력 텍스트: {input_text}")
-
-            # 1. 사용자 프로파일 로드
-            user_profile = self.get_user_profile(user_id)
-
-            # 2. 감정 분석
-            emotion = self.emotion_mapper.extract_emotion_from_text(input_text)
-            logger.info(
-                f"😊 감정 분석: V={emotion.valence:.3f}, A={emotion.arousal:.3f}, D={emotion.dominance:.3f}"
-            )
-
-            # 3. 프롬프트 생성
-            emotion_modifiers = self.emotion_mapper.emotion_to_prompt_modifiers(emotion)
-            personal_modifiers = user_profile.get_personalized_style_modifiers()
-
-            # 기본 프롬프트가 없으면 생성
-            if not base_prompt:
-                base_prompt = "digital art, beautiful scene"
-
-            final_prompt = f"{base_prompt}, {emotion_modifiers}, {personal_modifiers}"
-            final_prompt += ", high quality, detailed, masterpiece"
-
-            logger.info(f"🎯 최종 프롬프트: {final_prompt}")
-
-            # 4. 이미지 생성
-            if self.pipeline:
-                # SD 파이프라인 사용
-                with torch.autocast(
-                    self.device.type if self.device.type != "mps" else "cpu"
-                ):
-                    result = self.pipeline(
-                        prompt=final_prompt,
-                        num_inference_steps=num_inference_steps,
-                        guidance_scale=guidance_scale,
-                        width=width,
-                        height=height,
-                        output_type="pil",
-                    )
-
-                generated_image = result.images[0]
-                logger.info("✅ SD 파이프라인으로 이미지 생성 완료")
-            else:
-                # 폴백: 간단한 이미지 생성
-                generated_image = self._generate_fallback_image(emotion, width, height)
-                logger.info("⚠️ 폴백 이미지 생성기 사용")
-
-            # 5. 이미지 저장
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            image_filename = f"{user_id}_{timestamp}.png"
-            image_path = self.output_dir / image_filename
-            generated_image.save(image_path)
-
-            # 6. 데이터베이스에 기록
-            emotion_id = user_profile.add_emotion_record(
-                input_text=input_text,
-                emotion=emotion,
-                generated_prompt=final_prompt,
-                image_path=str(image_path),
-            )
-
-            # 7. 메타데이터 구성
-            metadata = {
-                "emotion_id": emotion_id,
-                "user_id": user_id,
-                "input_text": input_text,
-                "emotion": emotion.to_dict(),
-                "final_prompt": final_prompt,
-                "image_path": str(image_path),
-                "image_filename": image_filename,
-                "generation_params": {
-                    "num_inference_steps": num_inference_steps,
-                    "guidance_scale": guidance_scale,
-                    "width": width,
-                    "height": height,
-                },
-                "timestamp": timestamp,
-                "device": str(self.device),
-            }
-
-            logger.info(f"✅ 이미지 생성 완료: {image_path}")
-            return {"success": True, "image": generated_image, "metadata": metadata}
-
-        except Exception as e:
-            logger.error(f"❌ 이미지 생성 실패: {e}")
-            return {
-                "success": False,
-                "error": str(e),
-                "metadata": {"user_id": user_id, "input_text": input_text},
-            }
-
-    def _generate_fallback_image(
-        self, emotion: EmotionEmbedding, width: int = 512, height: int = 512
-    ) -> Image.Image:
-        """폴백 이미지 생성 (SD 파이프라인 실패시)"""
-
-        # 감정 기반 색상 생성
-        if emotion.valence > 0.3:
-            # 긍정적 감정 - 따뜻한 색상
-            base_color = [0.9, 0.8, 0.6]  # 따뜻한 노란색
-        elif emotion.valence < -0.3:
-            # 부정적 감정 - 차가운 색상
-            base_color = [0.6, 0.7, 0.9]  # 차가운 파란색
-        else:
-            # 중성 감정 - 중간 색상
-            base_color = [0.7, 0.7, 0.8]  # 회색빛
-
-        # 각성도 기반 강도 조정
-        intensity = 0.5 + abs(emotion.arousal) * 0.5
-        base_color = [c * intensity for c in base_color]
-
-        # 그라데이션 이미지 생성
-        image_array = np.zeros((height, width, 3))
-
-        for i in range(height):
-            for j in range(width):
-                # 중심에서의 거리 기반 그라데이션
-                center_x, center_y = width // 2, height // 2
-                distance = np.sqrt((j - center_x) ** 2 + (i - center_y) ** 2)
-                max_distance = np.sqrt(center_x**2 + center_y**2)
-
-                # 감정 기반 그라데이션 패턴
-                if emotion.dominance > 0:
-                    # 지배적 감정 - 중심에서 바깥으로
-                    factor = 1.0 - (distance / max_distance) * 0.5
-                else:
-                    # 수동적 감정 - 바깥에서 중심으로
-                    factor = 0.5 + (distance / max_distance) * 0.5
-
-                image_array[i, j] = [c * factor for c in base_color]
-
-        # numpy 배열을 PIL 이미지로 변환
-        image_array = np.clip(image_array * 255, 0, 255).astype(np.uint8)
-        return Image.fromarray(image_array)
-
-    def process_feedback(
-        self,
-        user_id: str,
-        emotion_id: int,
-        feedback_score: float,
-        feedback_type: str = "rating",
-        comments: str = None,
-        enable_training: bool = True,
-    ) -> Dict[str, Any]:
-        """사용자 피드백 처리 및 개인화 학습"""
-
-        try:
-            logger.info(f"📝 사용자 {user_id} 피드백 처리: 점수 {feedback_score}")
-
-            # 1. 사용자 프로파일 로드
-            user_profile = self.get_user_profile(user_id)
-
-            # 2. 피드백 저장
-            user_profile.add_feedback(
-                emotion_id=emotion_id,
-                feedback_score=feedback_score,
-                feedback_type=feedback_type,
-                comments=comments,
-            )
-
-            # 3. 강화학습 수행 (옵션)
-            training_result = None
-            if (
-                enable_training and self.trainer and feedback_score != 3.0
-            ):  # 중성 피드백 제외
-
-                # 해당 감정 기록 찾기
-                emotion_record = None
-                for record in user_profile.emotion_history:
-                    if record.get("id") == emotion_id:
-                        emotion_record = record
-                        break
-
-                if emotion_record:
-                    logger.info("🤖 개인화 학습 시작...")
-                    training_result = self.trainer.train_step(
-                        prompt=emotion_record["generated_prompt"],
-                        target_emotion=emotion_record["emotion"],
-                        user_profile=user_profile,
-                        num_inference_steps=8,  # 빠른 학습
-                    )
-                    logger.info(
-                        f"✅ 학습 완료: 보상 {training_result.get('total_reward', 0):.3f}"
-                    )
-
-            # 4. LoRA 어댑터 저장 (주기적)
-            if len(user_profile.feedback_history) % 5 == 0:  # 5번째 피드백마다 저장
-                self._save_user_lora_if_needed(user_id, user_profile)
-
-            # 5. 치료 인사이트 업데이트
-            insights = user_profile.get_therapeutic_insights()
-
-            result = {
-                "success": True,
-                "feedback_recorded": True,
-                "training_performed": training_result is not None,
-                "training_result": training_result,
-                "therapeutic_insights": insights,
-                "total_interactions": len(user_profile.emotion_history),
-                "total_feedbacks": len(user_profile.feedback_history),
-            }
-
-            logger.info("✅ 피드백 처리 완료")
-            return result
-
-        except Exception as e:
-            logger.error(f"❌ 피드백 처리 실패: {e}")
-            return {
-                "success": False,
-                "error": str(e),
-                "feedback_recorded": False,
-                "training_performed": False,
-            }
-
-    def _save_user_lora_if_needed(self, user_id: str, user_profile: UserEmotionProfile):
-        """필요시 사용자 LoRA 어댑터 저장"""
-        try:
-            if self.pipeline and hasattr(self.pipeline, "unet"):
-                # 현재 모델 상태를 LoRA로 저장
-                model_state = {
-                    "unet_state_dict": self.pipeline.unet.state_dict(),
-                    "user_preferences": user_profile.preference_weights,
-                    "training_metadata": user_profile.learning_metadata,
-                }
-
-                self.lora_manager.save_user_lora(user_id, model_state)
-                logger.info(f"💾 사용자 {user_id} LoRA 어댑터 저장")
-        except Exception as e:
-            logger.warning(f"⚠️ LoRA 저장 실패: {e}")
-
-    def get_user_insights(self, user_id: str) -> Dict[str, Any]:
-        """사용자 치료 인사이트 제공"""
-        user_profile = self.get_user_profile(user_id)
-        return user_profile.get_therapeutic_insights()
-
-    def get_emotion_history(self, user_id: str, limit: int = 10) -> List[Dict]:
-        """사용자 감정 히스토리 조회"""
-        user_profile = self.get_user_profile(user_id)
-        return user_profile.emotion_history[-limit:]
-
-    def cleanup_old_images(self, days_old: int = 30):
-        """오래된 이미지 파일 정리"""
-        try:
-            cutoff_date = datetime.now() - timedelta(days=days_old)
-            cleaned_count = 0
-
-            for image_file in self.output_dir.glob("*.png"):
-                if image_file.stat().st_mtime < cutoff_date.timestamp():
-                    image_file.unlink()
-                    cleaned_count += 1
-
-            logger.info(f"🧹 오래된 이미지 {cleaned_count}개 정리 완료")
-            return cleaned_count
-
-        except Exception as e:
-            logger.error(f"❌ 이미지 정리 실패: {e}")
-            return 0
-
-
-# =============================================================================
-# CLI 인터페이스
-# =============================================================================
+            print(f"  • {dir_name}/: 📁 생성 예정")
+
+    # 데이터베이스 상태
+    db_path = Path("user_profiles.db")
+    if db_path.exists():
+        size_mb = db_path.stat().st_size / 1024 / 1024
+        print(f"  • user_profiles.db: ✅ ({size_mb:.2f}MB)")
+    else:
+        print(f"  • user_profiles.db: 📄 생성 예정")
+
+    print("=" * 50)
+
+
+def show_usage_examples():
+    """사용 예시 표시"""
+
+    print("💡 사용 예시")
+    print("=" * 50)
+
+    examples = [
+        {
+            "title": "1. 기본 이미지 생성",
+            "command": 'python main.py --user-id "alice" --text "오늘 하루 정말 행복했다"',
+            "description": "사용자의 감정 일기를 분석하여 치료용 이미지 생성",
+        },
+        {
+            "title": "2. 상세 프롬프트와 함께 생성",
+            "command": 'python main.py --user-id "bob" --text "스트레스가 심하다" --prompt "평온한 자연 풍경"',
+            "description": "기본 프롬프트에 사용자 정의 프롬프트 추가",
+        },
+        {
+            "title": "3. 고품질 이미지 생성",
+            "command": 'python main.py --user-id "carol" --text "창의적인 기분" --steps 30 --guidance 9.0',
+            "description": "더 많은 스텝과 높은 가이던스로 고품질 이미지 생성",
+        },
+        {
+            "title": "4. 피드백 제공",
+            "command": 'python main.py --user-id "alice" --emotion-id 1 --feedback-score 4.5 --comments "정말 좋다"',
+            "description": "생성된 이미지에 대한 피드백 제공으로 개인화 학습",
+        },
+        {
+            "title": "5. 치료 인사이트 조회",
+            "command": 'python main.py --user-id "alice" --insights',
+            "description": "사용자의 감정 상태와 치료 진행 상황 확인",
+        },
+        {
+            "title": "6. 감정 히스토리 조회",
+            "command": 'python main.py --user-id "alice" --history 10',
+            "description": "최근 감정 기록과 생성된 이미지 히스토리 확인",
+        },
+        {
+            "title": "7. 시스템 정리",
+            "command": 'python main.py --user-id "admin" --cleanup 30',
+            "description": "30일 이상 된 이미지 파일 정리",
+        },
+    ]
+
+    for example in examples:
+        print(f"\n{example['title']}:")
+        print(f"  💡 {example['description']}")
+        print(f"  💻 {example['command']}")
+
+    print("\n" + "=" * 50)
 
 
 def main():
@@ -2232,21 +426,21 @@ def main():
             print("❓ 사용법:")
             print()
             print("1. 이미지 생성:")
-            print('   python script.py --user-id "alice" --text "오늘 기분이 좋다"')
+            print('   python main.py --user-id "alice" --text "오늘 기분이 좋다"')
             print()
             print("2. 피드백 제공:")
             print(
-                '   python script.py --user-id "alice" --emotion-id 1 --feedback-score 4.5'
+                '   python main.py --user-id "alice" --emotion-id 1 --feedback-score 4.5'
             )
             print()
             print("3. 치료 인사이트 조회:")
-            print('   python script.py --user-id "alice" --insights')
+            print('   python main.py --user-id "alice" --insights')
             print()
             print("4. 히스토리 조회:")
-            print('   python script.py --user-id "alice" --history 5')
+            print('   python main.py --user-id "alice" --history 5')
             print()
             print("5. 도움말:")
-            print("   python script.py --help")
+            print("   python main.py --help")
             print()
             print("💡 자세한 옵션은 --help를 참조하세요.")
 
@@ -2265,164 +459,6 @@ def main():
     print("✅ 작업 완료")
     return 0
 
-
-# =============================================================================
-# 시스템 정보 및 요구사항 체크
-# =============================================================================
-
-
-def check_system_requirements():
-    """시스템 요구사항 및 라이브러리 설치 상태 확인"""
-
-    print("🔍 시스템 요구사항 확인")
-    print("=" * 50)
-
-    # Python 버전
-    python_version = sys.version_info
-    print(
-        f"🐍 Python: {python_version.major}.{python_version.minor}.{python_version.micro}"
-    )
-
-    # 디바이스 정보
-    print(f"🔧 디바이스: {device} ({device_type})")
-
-    if device.type == "mps":
-        print("🍎 Apple Silicon 최적화 활성화")
-    elif device.type == "cuda":
-        print(f"🚀 CUDA 가능 (GPU: {torch.cuda.get_device_name()})")
-    else:
-        print("💻 CPU 모드")
-
-    # 메모리 정보
-    if device.type == "mps":
-        print("💾 통합 메모리 (Apple Silicon)")
-    elif device.type == "cuda":
-        gpu_memory = torch.cuda.get_device_properties(0).total_memory / 1024**3
-        print(f"💾 GPU 메모리: {gpu_memory:.1f}GB")
-
-    # 라이브러리 상태
-    print("\n📚 라이브러리 상태:")
-    libraries = {
-        "PyTorch": torch.__version__,
-        "Transformers": TRANSFORMERS_AVAILABLE,
-        "Diffusers": DIFFUSERS_AVAILABLE,
-        "PEFT": PEFT_AVAILABLE,
-        "PIL": True,  # 기본 라이브러리
-        "NumPy": np.__version__,
-    }
-
-    for lib, status in libraries.items():
-        if isinstance(status, bool):
-            status_str = "✅ 설치됨" if status else "❌ 미설치"
-        else:
-            status_str = f"✅ v{status}"
-        print(f"  • {lib}: {status_str}")
-
-    # 설치 권장사항
-    missing_libs = []
-    if not TRANSFORMERS_AVAILABLE:
-        missing_libs.append("transformers")
-    if not DIFFUSERS_AVAILABLE:
-        missing_libs.append("diffusers")
-    if not PEFT_AVAILABLE:
-        missing_libs.append("peft")
-
-    if missing_libs:
-        print(f"\n⚠️ 누락된 라이브러리: {', '.join(missing_libs)}")
-        print("설치 명령어:")
-        print(f"pip install {' '.join(missing_libs)}")
-    else:
-        print("\n✅ 모든 필수 라이브러리가 설치되어 있습니다!")
-
-    # 디렉토리 상태
-    print(f"\n📁 작업 디렉토리:")
-    dirs_to_check = ["generated_images", "user_loras"]
-    for dir_name in dirs_to_check:
-        dir_path = Path(dir_name)
-        if dir_path.exists():
-            file_count = len(list(dir_path.glob("*")))
-            print(f"  • {dir_name}/: ✅ ({file_count}개 파일)")
-        else:
-            print(f"  • {dir_name}/: 📁 생성 예정")
-
-    # 데이터베이스 상태
-    db_path = Path("user_profiles.db")
-    if db_path.exists():
-        size_mb = db_path.stat().st_size / 1024 / 1024
-        print(f"  • user_profiles.db: ✅ ({size_mb:.2f}MB)")
-    else:
-        print(f"  • user_profiles.db: 📄 생성 예정")
-
-    print("=" * 50)
-
-
-def show_usage_examples():
-    """사용 예시 표시"""
-
-    print("💡 사용 예시")
-    print("=" * 50)
-
-    examples = [
-        {
-            "title": "1. 기본 이미지 생성",
-            "command": 'python emotion_therapy.py --user-id "alice" --text "오늘 하루 정말 행복했다"',
-            "description": "사용자의 감정 일기를 분석하여 치료용 이미지 생성",
-        },
-        {
-            "title": "2. 상세 프롬프트와 함께 생성",
-            "command": 'python emotion_therapy.py --user-id "bob" --text "스트레스가 심하다" --prompt "평온한 자연 풍경"',
-            "description": "기본 프롬프트와 감정 분석을 결합한 맞춤형 이미지 생성",
-        },
-        {
-            "title": "3. 고품질 이미지 생성",
-            "command": 'python emotion_therapy.py --user-id "carol" --text "우울한 기분" --steps 25 --guidance 8.0',
-            "description": "더 많은 추론 스텝과 높은 가이던스로 고품질 이미지 생성",
-        },
-        {
-            "title": "4. 피드백 제공 (긍정적)",
-            "command": 'python emotion_therapy.py --user-id "alice" --emotion-id 1 --feedback-score 4.8 --comments "정말 마음에 든다"',
-            "description": "생성된 이미지에 대한 긍정적 피드백으로 개인화 학습",
-        },
-        {
-            "title": "5. 피드백 제공 (개선 필요)",
-            "command": 'python emotion_therapy.py --user-id "bob" --emotion-id 2 --feedback-score 2.3',
-            "description": "부정적 피드백을 통한 모델 개선",
-        },
-        {
-            "title": "6. 치료 진행도 확인",
-            "command": 'python emotion_therapy.py --user-id "alice" --insights',
-            "description": "감정 상태, 치료 진행도, 개인화 선호도 등 종합 분석",
-        },
-        {
-            "title": "7. 감정 히스토리 조회",
-            "command": 'python emotion_therapy.py --user-id "carol" --history 10',
-            "description": "최근 10개의 감정 기록과 생성된 이미지 이력 확인",
-        },
-        {
-            "title": "8. 시스템 정리",
-            "command": 'python emotion_therapy.py --user-id "admin" --cleanup 30',
-            "description": "30일 이상 된 이미지 파일들을 정리하여 저장 공간 확보",
-        },
-    ]
-
-    for example in examples:
-        print(f"\n{example['title']}")
-        print(f"💻 {example['command']}")
-        print(f"📝 {example['description']}")
-
-    print("\n" + "=" * 50)
-    print("🔧 고급 옵션:")
-    print("  --verbose          : 상세한 로그 출력")
-    print("  --no-training      : 피드백 시 학습 비활성화")
-    print("  --model MODEL_PATH : 사용할 Stable Diffusion 모델 지정")
-    print("  --width WIDTH      : 이미지 너비 (기본: 512)")
-    print("  --height HEIGHT    : 이미지 높이 (기본: 512)")
-    print("=" * 50)
-
-
-# =============================================================================
-# 메인 실행부
-# =============================================================================
 
 if __name__ == "__main__":
     # 시스템 정보 표시 (verbose 모드거나 도움말인 경우)
