+++ models/__init__.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+"""
+Emoseum Models - 감정 분석 및 사용자 프로필 모델
+"""
+
+from models.emotion import EmotionEmbedding
+from models.emotion_mapper import AdvancedEmotionMapper
+from models.user_profile import UserEmotionProfile
+from models.lora_manager import PersonalizedLoRAManager
+
+__all__ = [
+    "EmotionEmbedding",
+    "AdvancedEmotionMapper", 
+    "UserEmotionProfile",
+    "PersonalizedLoRAManager",
+]
