+++ training/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+"""
+Emoseum Training - 강화학습 훈련 모듈
+"""
+
+__all__ = []
