+++ utils/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+"""
+Emoseum Utils - 유틸리티 모듈
+"""
+
+__all__ = []
